Swap the places of ranger and the shell
Closes #4.

I originally intended the shell to be below ranger but removed it in
744ccd7 for the sake of compatibility. This new solution should work
with tmux 1.8+.

The previous behavior may be restored by setting
TEXAS_CONFIG_NOSWAP=1.

It seems in tmux 1.8 shell is being focused on start. In tmux 2.1
ranger is being focused, as intended. Currently WONTFIX unless it's
indeed an issue.